web/backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-filter==25.1
django-resized==1.0.3
djangorestframework==3.16.1
idna==3.10
Markdown==3.8.2
pillow==11.3.0
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
